feat: add settings page with sections
