+++ app/src/renderer/layout/common/main-sidebar/navbar/SettingsMenu.tsx
@@ -1,59 +1,15 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { Alert, Box, Divider, Drawer, IconButton, Link, MenuItem, Stack, TextField, Typography } from '@mui/material';
-import { useSettings } from 'renderer/contexts/SettingsContext';
-import { FormattedMessage } from 'react-intl';
-import { CloseOutlined } from '@mui/icons-material';
-import { alpha } from '@mui/material/styles';
-import PerfectScrollbar from 'react-perfect-scrollbar';
-import { COMPONENTS_SPACING, NAVBAR_HEIGHT } from 'renderer/constants/ui';
-import { ThemeSource } from 'infra/ui/models/electronTheme';
+import React, { useCallback, useEffect } from 'react';
 import { useSubscribeToEvent } from 'renderer/apis/requests/subscriptions/subscribeToEvent';
 import { IpcRendererEvent } from 'electron';
-import { useRestartApp } from 'renderer/apis/requests/ui/restartApp';
-import { useAppUpdates } from 'renderer/contexts/AppUpdatesContext';
-import { useSnackbar } from 'notistack';
-import { useAnalytics } from 'renderer/contexts/AnalyticsContext';
-import semverGt from 'semver/functions/gt';
 import NavbarIconButton from './NavbarIconButton';
+import { useNavigate } from 'react-router-dom';
+import { urls } from 'renderer/routes/urls';
 
 export default function SettingsMenu() {
-  const {
-    themeSource,
-    setThemeSource,
-    notificationsActive,
-    setNotificationsActive,
-    notificationsSoundActive,
-    setNotificationsSoundActive,
-    analyticsEnabled,
-    setAnalyticsEnabled,
-    errorReportingEnabled,
-    errorReportingChanged,
-    setErrorReportingEnabled,
-  } = useSettings();
-  const {
-    autoUpdateSupported,
-    autoUpdateEnabled,
-    setAutoUpdateEnabled,
-    newVersionInfo,
-    newVersionDownloaded,
-    checkForUpdates,
-    downloadUpdate,
-    installUpdate,
-  } = useAppUpdates();
-  const { enqueueSnackbar } = useSnackbar();
-  const { track } = useAnalytics();
+  const navigate = useNavigate();
 
-  const [open, setOpen] = useState<boolean>(false);
-
-  const toggleOpenHandler = useCallback((): void => {
-    setOpen((prev) => !prev);
-  }, []);
-
-  const restartAppState = useRestartApp();
-
-  const restartAppHandler = useCallback((event: React.MouseEvent): void => {
-    event.preventDefault();
-    restartAppState.mutate({});
+  const openSettingsHandler = useCallback((): void => {
+    navigate(urls.settings.url);
   }, []);
 
   const subscribeToTriggerEventsState = useSubscribeToEvent();
@@ -64,7 +20,7 @@ export default function SettingsMenu() {
       unsubscribe = await subscribeToTriggerEventsState.mutateAsync({
         event: 'trigger:openSettings',
         listener: (event: IpcRendererEvent) => {
-          setOpen(true);
+          openSettingsHandler();
         },
       });
     })();
@@ -73,294 +29,5 @@ export default function SettingsMenu() {
     };
   }, []);
 
-  const [appVersion, setAppVersion] = useState<string>('');
-
-  useEffect(() => {
-    (async () => {
-      setAppVersion(await window.ui.getAppVersion());
-    })();
-  }, []);
-
-  const appUpdatesView = useMemo<'none' | 'check' | 'download' | 'install'>(() => {
-    if (!appVersion) {
-      return 'none';
-    }
-    if (newVersionDownloaded && semverGt(newVersionDownloaded.version, appVersion) && autoUpdateSupported) {
-      return 'install';
-    }
-    if (newVersionInfo && semverGt(newVersionInfo.version, appVersion)) {
-      return 'download';
-    }
-    return 'check';
-  }, [appVersion, autoUpdateSupported, newVersionInfo, newVersionDownloaded]);
-
-  const checkForUpdatesHandler = useCallback(
-    async (event: React.MouseEvent): Promise<void> => {
-      event.preventDefault();
-
-      track({ name: 'settings_check_for_updates' });
-
-      const result = await checkForUpdates();
-      if (result) {
-        enqueueSnackbar(<FormattedMessage id="newVersionIsAvailable" values={{ version: result.version }} />, {
-          variant: 'info',
-        });
-      } else {
-        enqueueSnackbar(<FormattedMessage id="appIsUpToDate" />, { variant: 'info' });
-      }
-    },
-    [checkForUpdates]
-  );
-
-  const downloadUpdateHandler = useCallback(
-    (event: React.MouseEvent): void => {
-      event.preventDefault();
-
-      track({ name: 'settings_download_update' });
-
-      const downloadType = downloadUpdate();
-      if (downloadType === 'internal') {
-        enqueueSnackbar(<FormattedMessage id="downloadStarted" />, { variant: 'info' });
-      }
-    },
-    [downloadUpdate]
-  );
-
-  const installUpdateHandler = useCallback(
-    (event: React.MouseEvent): void => {
-      event.preventDefault();
-
-      track({ name: 'settings_install_update' });
-
-      installUpdate();
-    },
-    [installUpdate]
-  );
-
-  const canOpenOsSettings = useMemo<boolean>(() => window.notifications.canOpenOsSettings(), []);
-
-  const openOsSettingsHandler = useCallback((event: React.MouseEvent): void => {
-    event.preventDefault();
-    window.notifications.openOsSettings();
-  }, []);
-
-  return (
-    <>
-      <NavbarIconButton titleId={'settings'} icon={'SettingsOutlined'} onClick={toggleOpenHandler} />
-
-      <Drawer
-        anchor={'right'}
-        open={open}
-        onClose={toggleOpenHandler}
-        PaperProps={{
-          sx: {
-            width: '100%',
-            maxWidth: 320,
-            backdropFilter: 'blur(6px)',
-            WebkitBackdropFilter: 'blur(6px)', // Fix on Mobile
-            backgroundColor: (theme) => alpha(theme.palette.background.paper, 0.72),
-            boxShadow: (theme) => theme.customShadows.z24,
-          },
-        }}
-        BackdropProps={{ invisible: true }}
-      >
-        <Box sx={{ display: 'flex', flexDirection: 'column' }}>
-          <Stack
-            direction="row"
-            alignItems="center"
-            justifyContent="space-between"
-            sx={{ height: NAVBAR_HEIGHT - 1, px: COMPONENTS_SPACING }}
-          >
-            <Typography variant="subtitle2">
-              <FormattedMessage id={'settings'} />
-            </Typography>
-            <IconButton size={'small'} onClick={toggleOpenHandler}>
-              <CloseOutlined fontSize={'small'} />
-            </IconButton>
-          </Stack>
-          <Divider />
-          <Box sx={{ flexGrow: 1 }}>
-            <PerfectScrollbar>
-              <Stack spacing={COMPONENTS_SPACING} sx={{ p: COMPONENTS_SPACING }}>
-                <TextField
-                  fullWidth
-                  label={<FormattedMessage id="theme" />}
-                  margin="normal"
-                  select
-                  value={themeSource}
-                  onChange={(e) => setThemeSource(e.target.value as ThemeSource)}
-                >
-                  <MenuItem value={'system'}>
-                    <FormattedMessage id="system" />
-                  </MenuItem>
-                  <MenuItem value={'dark'}>
-                    <FormattedMessage id="dark" />
-                  </MenuItem>
-                  <MenuItem value={'light'}>
-                    <FormattedMessage id="light" />
-                  </MenuItem>
-                </TextField>
-
-                <TextField
-                  fullWidth
-                  label={<FormattedMessage id="notifications" />}
-                  margin="normal"
-                  select
-                  value={notificationsActive}
-                  onChange={(e) => setNotificationsActive(e.target.value === 'true')}
-                >
-                  <MenuItem value={'true'}>
-                    <FormattedMessage id="yes" />
-                  </MenuItem>
-                  <MenuItem value={'false'}>
-                    <FormattedMessage id="no" />
-                  </MenuItem>
-                </TextField>
-
-                {notificationsActive && (
-                  <>
-                    <Alert severity={'info'} variant={'outlined'}>
-                      <FormattedMessage
-                        id="notReceivingNotifications"
-                        values={{
-                          settings: canOpenOsSettings ? (
-                            <Link href={`#`} onClick={openOsSettingsHandler} sx={{ color: 'info.main' }}>
-                              <FormattedMessage id={'operatingSystemSettings'} />
-                            </Link>
-                          ) : (
-                            <FormattedMessage id={'operatingSystemSettings'} />
-                          ),
-                        }}
-                      />
-                    </Alert>
-
-                    <TextField
-                      fullWidth
-                      label={<FormattedMessage id="notificationsSound" />}
-                      margin="normal"
-                      select
-                      value={notificationsSoundActive}
-                      onChange={(e) => setNotificationsSoundActive(e.target.value === 'true')}
-                    >
-                      <MenuItem value={'true'}>
-                        <FormattedMessage id="yes" />
-                      </MenuItem>
-                      <MenuItem value={'false'}>
-                        <FormattedMessage id="no" />
-                      </MenuItem>
-                    </TextField>
-                  </>
-                )}
-
-                <TextField
-                  fullWidth
-                  label={<FormattedMessage id="allowTracking" />}
-                  margin="normal"
-                  select
-                  value={analyticsEnabled}
-                  onChange={(e) => setAnalyticsEnabled(e.target.value === 'true')}
-                  helperText={<FormattedMessage id="allowTrackingDescription" />}
-                >
-                  <MenuItem value={'true'}>
-                    <FormattedMessage id="yes" />
-                  </MenuItem>
-                  <MenuItem value={'false'}>
-                    <FormattedMessage id="no" />
-                  </MenuItem>
-                </TextField>
-
-                <TextField
-                  fullWidth
-                  label={<FormattedMessage id="allowErrorReporting" />}
-                  margin="normal"
-                  select
-                  value={errorReportingEnabled}
-                  onChange={(e) => setErrorReportingEnabled(e.target.value === 'true')}
-                  helperText={<FormattedMessage id="allowErrorReportingDescription" />}
-                >
-                  <MenuItem value={'true'}>
-                    <FormattedMessage id="yes" />
-                  </MenuItem>
-                  <MenuItem value={'false'}>
-                    <FormattedMessage id="no" />
-                  </MenuItem>
-                </TextField>
-
-                {errorReportingChanged && (
-                  <Alert severity={'info'} variant={'outlined'}>
-                    <FormattedMessage
-                      id="restartAppChangesTakeAffect"
-                      values={{
-                        restart: (
-                          <Link href={'#'} color={'info.main'} onClick={restartAppHandler}>
-                            <FormattedMessage id={'restart'} />
-                          </Link>
-                        ),
-                      }}
-                    />
-                  </Alert>
-                )}
-
-                {autoUpdateSupported && (
-                  <TextField
-                    fullWidth
-                    label={<FormattedMessage id="automaticUpdates" />}
-                    margin="normal"
-                    select
-                    value={autoUpdateEnabled}
-                    onChange={(e) => setAutoUpdateEnabled(e.target.value === 'true')}
-                  >
-                    <MenuItem value={'true'}>
-                      <FormattedMessage id="yes" />
-                    </MenuItem>
-                    <MenuItem value={'false'}>
-                      <FormattedMessage id="no" />
-                    </MenuItem>
-                  </TextField>
-                )}
-
-                <TextField
-                  fullWidth
-                  label={<FormattedMessage id="appVersion" />}
-                  margin="normal"
-                  value={appVersion}
-                  inputProps={{ readOnly: true }}
-                  helperText={
-                    <>
-                      {appUpdatesView === 'check' && (
-                        <Link href={`#`} onClick={checkForUpdatesHandler} variant={'inherit'}>
-                          <FormattedMessage id={'checkForUpdates'} />
-                        </Link>
-                      )}
-                      {appUpdatesView === 'download' && (
-                        <>
-                          <FormattedMessage
-                            id="newVersionIsAvailableAndReady"
-                            values={{ version: newVersionInfo?.version }}
-                          />
-                          <br />
-                          <Link href={`#`} onClick={downloadUpdateHandler} variant={'inherit'}>
-                            <FormattedMessage id={'downloadUpdate'} />
-                          </Link>
-                        </>
-                      )}
-                      {appUpdatesView === 'install' && (
-                        <>
-                          <FormattedMessage id="appUpdateDownloadedAndReady" />
-                          <br />
-                          <Link href={`#`} onClick={installUpdateHandler} variant={'inherit'}>
-                            <FormattedMessage id={'quitAndInstall'} />
-                          </Link>
-                        </>
-                      )}
-                    </>
-                  }
-                />
-              </Stack>
-            </PerfectScrollbar>
-          </Box>
-        </Box>
-      </Drawer>
-    </>
-  );
+  return <NavbarIconButton titleId={'settings'} icon={'SettingsOutlined'} onClick={openSettingsHandler} />;
 }
